*: Convert proto_redistnum to return a uint8_t
There are places that we are playing fast and loose
with signed -vs- unsigned integers.  Let's convert
proto_redistnum to only return a uint8_t.  Find
and fix all the places that matter.  This was
found by coverity.

(cherry picked from commit 755a879be512e16800da2096bbd24088e4de5dd1)

# Conflicts:
#	babeld/babel_zebra.c

+++ lib/log.c
@@ -544,7 +544,7 @@ int proto_name2num(const char *s)
 	return -1;
 }
 
-int proto_redistnum(int afi, const char *s)
+uint8_t proto_redistnum(int afi, const char *s)
 {
 	if (!s)
 		return ZEBRA_ROUTE_ERROR;
